requirements.txt
flask
pywebview
PyQt5
pyinstaller
PyQtWebEngine
